pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sec-semantic-search"
version = "0.1.0"
description = "A semantic search system for SEC filings (10-K, 10-Q)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
authors = [
    { name = "Eren Yanic" },
]
keywords = ["sec", "edgar", "semantic-search", "embeddings", "chromadb"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    # SEC filing retrieval
    "edgartools>=5.13.0",

    # HTML parsing
    "doc2dict>=0.6.9",

    # Embeddings and ML
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",

    # Vector database
    "chromadb>=1.0.0",

    # Web interface (Streamlit — legacy, kept for reference)
    "streamlit>=1.30.0",

    # REST API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",

    # CLI framework
    "typer>=0.12.0",
    "rich>=13.0.0",

    # Configuration
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
sec-search = "sec_semantic_search.cli.main:app"
deprecated-sec-search-web = "sec_semantic_search.web_deprecated.run:main"
sec-search-api = "sec_semantic_search.api.run:main"

[project.urls]
Homepage = "https://github.com/ErenYanic"
Repository = "https://github.com/ErenYanic/SEC-SemanticSearch"

[tool.hatch.build.targets.wheel]
packages = ["src/sec_semantic_search"]

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["sec_semantic_search"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
